requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0

langchain==0.2.6
langchain-community==0.2.6
langchain-openai==0.1.17
langgraph==0.1.19

openai>=1.32.0,<2.0.0
tiktoken

faiss-cpu
python-dotenv
langfuse==2.33.0
